refactor: collapse plugin barrel into plugin/index.ts
